+++ app/models/pay/charge.rb
@@ -17,6 +17,7 @@ module Pay
 
     # Store the payment method kind (card, paypal, etc)
     store_accessor :data, :paddle_receipt_url
+    store_accessor :data, :stripe_receipt_url
     store_accessor :data, :stripe_account
 
     # Payment method attributes
@@ -29,6 +30,9 @@ module Pay
     store_accessor :data, :username # Venmo
     store_accessor :data, :bank
 
+    store_accessor :data, :amount_captured
+    store_accessor :data, :invoice_id
+    store_accessor :data, :payment_intent_id
     store_accessor :data, :period_start
     store_accessor :data, :period_end
     store_accessor :data, :line_items
@@ -47,6 +51,8 @@ module Pay
       scope processor_name, -> { where(processor: processor_name) }
     end
 
+    delegate :capture, :credit_note!, :credit_notes, to: :payment_processor
+
     def self.find_by_processor_and_id(processor, processor_id)
       joins(:customer).find_by(processor_id: processor_id, pay_customers: {processor: processor})
     end
@@ -61,6 +67,10 @@ module Pay
 
     def processor_charge
       payment_processor.charge
+    end
+
+    def captured?
+      amount_captured > 0
     end
 
     def refund!(refund_amount = nil)