+++ lib/pay/braintree/webhooks/subscription_canceled.rb
@@ -11,7 +11,12 @@ module Pay
           pay_subscription = Pay::Subscription.find_by_processor_and_id(:braintree, subscription.id)
           return unless pay_subscription.present?
 
-          pay_subscription.update!(ends_at: Time.current, status: :canceled)
+          ends_at = Time.current
+          pay_subscription.update!(
+            status: :canceled,
+            trial_ends_at: (ends_at if pay_subscription.trial_ends_at?),
+            ends_at: ends_at
+          )
         end
       end
     end